create generator.py to build commands
The script can be used to create a random graph of actors and movies.
Of course, neither the actors nor the movies are relevant, they are just
numbers, but the graph can be used to build and play with a database.

You can select the number of actors you want in the graph, the maximum
number of movies (it can be smaller depending on other parameters), the
average/standard deviation number of roles per actors.