add parse_actors.py to import IMDb lists
parse_actors.py can be used to import IMDb list files and convert them
into the same kind of commands that was generated by generator.py.

The list are not free but can be used under certain circumstances.
You will have to download it yourself if you want to use them. You will
find more information on the following page:

    http://www.imdb.com/interfaces

This script is parsing the files actors.list and actresses.list to build
the graph commands. You can use it like this:

    python parse_actors.py actors.list actresses.list