create importer.c to build graph database
The importer is used to convert from graph commands and build a graph
database. The database is a own-made database with index for actors, and
fast retrieval of relations between actors and movies (called roles).

It uses glib data structures (HashTables, and Red-Black Binary tree) to
build the database as fast as possible.

It takes a single parameter that is the file in which it will save the
database. It reads from standard input the graph commands.

It can be used like this (using generator):

    python generator.py | ./importer graph.db

Or using IMDb files:

    python parse_actors.py actors.list actresses.list |
    ./importer graph.db