Create bacon.c to look in the graph
bacon program is used to look into the graph and find the shortest path
between two actors. It will output the path as a dot file.

It can take many parameters to select which type of "show" can be
selected to create the path; and also take a couple of actors.

It can be used like this:

    ./bacon graph.db "De Niro, Robert" "Bacon, Kevin (I)" >graph.dot

Currently, the names have to be given exactly as they were entered when
building the database (only exact match). Using IMDb file, it means you
may have to add something like "(I)" to disambiguate the names.

pack.c and pack.h are used to manipulate the database file, and drive
this logic away from the shortest_path finding algorithm.